#18 Unele legi nu sunt identificate corect. DOM parser.
Changed the s02_votes_get.py script to use a DOM parser instead of
regex. Regex parsers fail when one link is broken and jump to the next
creating an offset. The DOM parser reads table rows and if one link is
broken both votes and links hrefs will be treated the same.